WarpImages: support resampling warped images.
This commit adds the capability of inverting an affine transform and then resampling the voxel intensities of one
(un-warped) image array onto a second (geometrical placholder) image array. Transferred voxel intensities are linearly
interpolated on the un-warped image array by means of the inverse transform.

This commit is slow and a bit awkward to use. It seems to work well, however the method will not easily generalize to
other transformation techniques, and there are faster ways to implement the affine transform specifically. So this
implementation is likely to change in the future.