Add TabulateImageMetadata operation.
This operation functions similar to the DumpImageMetadataOccurrencesToFile operation, but decouples data extraction from
writing to file. Results are stored as a table which can be manipulated or used in later operations.

TabulateImageMetadata also support both 'wide' and 'tall' tables to handle scenarios where output from multiple passes
can be more consistently combined together in a normal spreadsheet arrangement ('wide') and a two-column key-value style
that can be easily grepped ('tall').