Add operation for decomposing images using SVD.
The outputs are (1) a set of basis images, and (2) the corresponding singular values imbued as metadata in the basis
image.

This operation can, in principle, be used to perform SVD-based (whole-image) classification. All lnecessary data is
present. However, there is currently no operation exposed for this purpose. Classification may be added at a later time.