FITS: allow img metadata to take priority when loading.
This commit adds support for recently-added upstream Ygor support for planar_image key-value metadata.
Previously, no metadata was embedded in FITS images, so generic metadata was added to each image.
This commit permits any existing metadata to take priority over generic metadata, but still keeps the
generic metadata mechanism around to fill in the blanks in case metadata is not present. A complicating
factor is that the metadata is 'evolved' between images to keep them linked but not identical; any
existing metadata is used as a base for evolution, but existing metadata takes priority over evolved
metadata.