SDL_Viewer: support ROI contour editing.
This commit adds support for editing existing contours. However, rathan than destructive editing, only copies of
existing contours can be edited. To support use-case of replacing an existing ROI with an edited version, the
ability to overwrite an existing ROI is also provided -- but has to be explicitly enabled by the user to prevent
accidental data loss. Attempting to overwrite an existing contour without selecting the overwrite option causes the save
to fail and emits a warning.