SDL_Viewer: promote aliases in operations list.
This commit 'promotes' aliases to be on-par with canonical function names for the purposes of interactively generating
scripts. Previously, aliases were listed but not easily discoverable since they were not in alphabetical order. This
commit adds all aliases as if they were canonical operations so they appear in alphabetical order and their full
documentation can be viewed.