Add support for a non-DICOM contour collection (ROI) file format.
This commit adds support for a custom plaintext format for contour collections. The aim is to provide a simplified
alternative to DICOM which is easier to modify with basic tooling.

Both read and write support are provided in this commit. Test contours successfully round-trip, but no test cases have
been added yet.