Contour_Collection_File_Loader: increase parser leniency.
This commit allows comments to have any amount of preceding whitespace before the '#' character, but comments are still
only possible on otherwise empty lines.

This commit also allows skipping the 'contour_start' and 'collection_start' statements, which allow metadata to
fallthrough for all contours in a collection. The emitter currently does not support this because contours could, in
principle, all have variable and inconsistent metadata. However, this dramatically reduces the amount of boilerplate
required to specify contour collections with homogeneous metadata.