SDL_Viewer: bug fix: drawing/contouring brush scale.
This commit fixes a minor bug possible with non-square images. During drawing and countouring, a brush overlay shape is
displayed. Previously, the brush overlay scale could be invalid when displaying non-square images. Note that the actual
brush stroke size and position were correct, but the overlay shape showing where the brush was hovering had an incorrect
scale.

The problem was inconsistent naming between upstream Ygor and upstream ImGui, which use 'x' to refer to different
things. This commit fixes the scale issue.