Write_CT_Images: emit LUT and converted pixels.
Previously, CT pixels were rounded and exported as-is, with an identity LUT transformation. This often works for CT,
since the pixel intensities are normally integer-valued and well within the (unscaled) 16-bit range.

This commit adds support for reading (any modality that embeds a simple linear transformation) and writing (CT) of
linear floating point LUT transformations.