Write_CT_Images: prefer identity transform when possible.
This commit adds a (slow) scan to determine whether pixels need to be transformed when compressing to 16-bit integer. If
all pixels intensities are already suitable for direct conversion, an identity transformation is used. This results in
the most retention of precision, but also permits floating point numbers to be encoded when needed.