SDL_Viewer: auto-scroll logs.
This commit also tries to ensure logs stay visible during operations, which is when they are most useful.
Additionally, the style of text display was switched from a text input area to a static text block. The latter does not
directly support copying, but it does support better scrolling integration.