SDL_Viewer: avoid tray notification cascade.
SDL_Viewer now relays YLOGWARN and higher severity messages to a tray notification, to better communicate when something
goes wrong. If the tray notification fails to be emitted, currently a YLOGWARN can be emitted. Previously this could
cause a loop of continously failing notifications and re-relays.

This commit adds a switch that automatically disables tray notifications if one cannot be emitted. This may result in
missing notifications, which is not ideal, but it avoids re-relays and the log can still be pulled up for viewing (so no
messages will be lost). The default setup also logs to the terminal, so messages should still be received in duplicate
if tray notifications are disabled.