Add Tetrominoes operation.
This operation implements a simple tetromino game using images. It is possible to play this game on arbitrary images,
however only 2D images are implemented. Each single image of an image array will be simulated independently.

This operation requires manual input to alter the placement of the tetrominoes. Automated placement is not (yet?)
supported.