Replace ASIO thread pool with simplistic thread pool.
The replacement uses only C++11 threading primitives. It also eliminates warnings produced by ThreadSanitizer and
Valgrind's drd tool involving a condition variable that is not protected by a mutex when notifying somewhere inside
ASIO.

Note that the ASIO package is still listed as a dependency and will continue to be installed in header-only form.
(I'm still evaluating whether to use ASIO for networking purposes later.)