SDL_Viewer: attempt fix for contouring lock-up on Windows.
On Linux there is a data race detected using valgrind and TSAN, but everything works fine. On Windows, saving contours
sometimes causes a lock-up when performing the same actions. From what I can tell, this appears to be due to assignment
of a std::future variable using std::async(). It's extremely difficult to debug, so hard to tell.

The test case added here both demonstrates the same issue using std::future, and absence of the issue when using
std:::shared_future. The key appears to be assigning the future object when an existing future is held on the LHS of the
assignment. There is a well-known limitation of std::async() futures blocking in their destructor, and I wonder if the
assignment mid-task causes the running code to write to the old, overwritten memory location before 'noticing' the task
has been cancelled. If so, it seems strange to allow assignment of a running std::future. However, I'm not 100% certain
this is the actual problem. So this commit might require additional follow-up work.