RPC: add test for class alignment.
This commit Thrift-based serialization/deserialization for point_set<double>. It also adds a simple sizeof() check to
detect at compile time if the point_set class members were altered (= members were added or deleted, but not moved or
type was changed).

This specific check disregards padding bytes, since on the test machine padding is not needed. CI will determine if I
can rely on this for all current cross-ompilation targets.