RPC: attempt to deal with armhf/armv7 class alignment.
CI fails only for armhf and armv7 due to class alignment differing compared with x86 and x86_64 alignment.
This commit adds a slight tweak that accounts for class alignment, but is unlikely to work.

Wish I could print the sizeof() at compile time somehow. Failed static_assert() on these systems do not
seem to print the failed condition, which is unfortunate for debugging the size discrepancy.