SDL_Viewer: hoist contour extractor thread closure.
This commit converts the mutable contour extractor closure to a const closure by hoisting it up earlier in the
SDL_Viewer operation.

Despite refactoring the contour extractor, and not being able to detect any races in TSAN while testing on Linux, I'm
still seeing lock-ups on MXE builds. Other uses of std::async/std::shared_future do not seem to exhibit this behaviour,
and they do not use mutable closures. So this commit is an attempt to mitigate lock-ups due to some unknown
thread-related issue only occuring on Windows.