CMake: back-out change that will require larger refactor.
The previous commit added more explicit linking for dialogshim and imebrashim libraries. However, it caused definitions
of some symbols to appear in more than one object (i.e., an 'o' file and a 'so' file) when compiling executables, which
causes linking to fail on most systems. This will require refactoring the linking strategy so will be reverted for now.