Notifications: adjust handler presence test logic.
When probing to see if zenity is available on Windows, the check logic appears to close the console window. It seems
that the previous shell command's "exit" propagated to the console window rather than running in a separate window.

This commit removes the "exit" call and simplifies the logic for testing if a command is available.