Notifications: stop console disappearing on Windows.
The '-WindowStyle hidden' option sent to powershell causes the attached console to disappear. Omitting the option fixes
the problem, but still leaves the console accessible, which could cause similar issues in the future (e.g., if
powershell behaviour changes, or the syntax changes and causes the powershell process to disable the console on
termination).

An additional prefix wrapper (start /B /wait) was added to attempt to create a separate console for the powershell
invocation, but it seems ineffectual for some reason specifically when using powershell. I will leave it in case it
helps isolate the console from future messages/formatting causes by the powershell process.