Documentation: expose per-operation documentation.
This is likely more convenient than the full reference manual, especially since each operation's documentation can be
opened in a separate window.

Better documentation of operation aliases is provide with the canonical operation now explicitly listing aliases in the
reference manual. er-operation documentation when using an alias refers to the canonical name, and the canonical name
lists all aliases. There is a risk this might be confusing, but nevertheless things should always just work in terms of
scripts, and the documentation for aliases can be explicitly looked up by the user now.