Fork: provide fork-equivalent on Windows.
The 'equivalent' fork() is based on a detached std::thread, which barely achieves the core goal of fork (i.e., create a
second thread of execution in a disconnected memory space so that state changes do not propagate to the parent).

There are serious limitations due to sharing the same process (e.g., termination affects all "forks"; internal
state is shared and non-thread-local globals are likely inadvertantly shared; signals are shared) and the relevant
application state is manually copied before the fork occurs, so it will be extremely slow compared to a proper
copy-on-write fork().

The specifics of Windows fork emulation are subject to change.