SDL_Viewer: support 3-channel RGB viewing.
This commit adds support to view the first three channels of an image as 8-bit RGB coordinates. This option only appears
when the image has three or more channels, and resets automatically when an image is viewed that does not have enough
channels. Window and leveling is not applied, and the colour map is also not used. Despite these limitations, this
functionality provides a more natural way to view common 3- and 4-channel RGB image formats, which can be helpful.