MXE: workaround use of 'debian:stable' in existing Docker image.
The existing DCMA Docker image used for MXE cross-building was based on debian:stable, which now resolves to
debian:oldstable. Attempting to run `apt` inside the image prodices an error since the codename that previously referred
to bullseye now refers to bookworm. Replacing 'stable' in the sources.list file with 'bullseye' (which is the current
'oldstable') ameliorates the issue.

There are two times where this change needs to occur: when building the Docker image, and when using it for builds/CI.
Going forward, when a new MXE Docker image will be created, this change will be made explicitly. For now, we rely on
editing sources.list on-the-fly in a way that will (hopefully) be idempotent.