WIP: add Isolate operation.
This operation selects a portion of the full state as a view, and then invokes children operations on the view. The view
appears as a complete Drover object to the children, and un-selected objects are not accessible from the view.
Afterward, the view is destroyed but any deletions/creations are added back to the Drover.