Add CompileScript operation to explicitly validate/compile/run scripts at runtime.
This complements the existing SDL_Viewer script editor and DCMA script file loader, but can be invoked from the
command-line (or within a script) to provide non-interactive runtime loading and execution of a script. This is also
useful for validating/linting scripts from the command line, acting as a compiler.

Note that, at least for now, nothing is emitted when compiling; the outputs are stored in-memory only. There is
currently no portability guaranteed for DCMA scripts, especially since some instances may be missing specific
operations and/or functionality. At least for now, scripts are more portable.