Scripts: make variable assignment flow from top to bottom.
Previously, redefining a variable would cause retrospective use of the variable to hold the redefined value. For
example,

    |  x = 123;
    |  noop(y = x);
    |  x = 456;

would invoke noop() with y = 456. This behaviour was documented (and a warning was issued during validation for each
definition). However, it was still surprising, and I can't think of situations where I would specifically want this
behaviour.

This commit ensures variable updates never propagate 'backwards', so in this case noop() would be invoked with y = 123.