WIP: SDL_Viewer: add basic support for interactive guides.
This commit adds an extremely basic window setup for interactive guides that walk the user through a tour, tutorial, or
some complex procedure. At the moment, only text can be rendered for each step/stage of the guide.

Note: this commit adds the core guide functionality, but it is not exposed to the user; there is no way to add or open
guides yet.