WIP: SDL_Viewer: support base64-encoded images.
This commit adds support for decoding and displaying images embedded in the guide. Images should be encoded as base64,
but otherwise can be one of the 'common' 8-bit image formats (e.g., png, jpg).

Note that this code is still not accessible to users. There is currently no way to load or view a guide (yet).