AnalyzeHistograms: add special case for V(0 Gy) constraints.
The previous (general) case performed an interpolation at the boundary of the line sample, which is an extrapolation and
was thus forced to zero. This commit does not 100% address the issue of very low doses causing the line sample to
extrapolate (and thus force to zero), but it does provide a workaround for the most typical case in which this situation
would arise (i.e., the user would like to compare the entire volume).