SDL_Viewer: support cell selection.
This commit adds support for selecting multiple cells using 'typical' shortcuts (ctrl for multiple, shift for
rectangular). Supporting selection required replacing the existing grid of InputText with Selectable for non-editing
cells and InputText for the cell being edited.

Tab and Enter movements require a debounce that requires multiple frames to work, which is a bit unsatisfying.