SDL_Viewer: tables: support pasting contents.
Pasting cell content required some sort of indicator for the user to show *where* the cells would be pasted, so this
commit also adds a visual indicator to the most recently clicked cell.

Note that the SDL function is directly called for accessing the clipboard because the imgui method did not immediately
call SDL_free().