SDL_Viewer: Tables: add keyboard adjacent cell navigation.
This commit reverts back to Selectable from InvisibleButton in order to avoid a mouse clikc dead-space between each
cell. While Selectable avoids this dead-space, it adds annoying keyboard navigation that seemingly cannot be disabled.

To work around this, the navigation focused cell is hidden by making the style have zero alpha and manually scrolling
when a new cell is selected. This can result in a frame of scroll in the wrong direction, which appears as jitter when
holding down a keyboard navigation button. Ultimately it seems to work.

I also support entering text editing mode when a keypress is detected. However, currently the first keypress is ignored.
There is roughed-in support to explicitly add the keypress to the cell, but it interferes with the InputText editing
features. Specifically the "highlight the existing text upon entering edit mode" and "press escape to revert contents"
features. I haven't figured out a reasonable workaround for this. Imgui encodes typed characters as wide unicode format,
but does not provide a way to decode back to utf8 -- there is a routine specifically for this, but for some reason it is
not exposed in the public API. I have no idea what the rationale for this is, and it is even marked as FIXME in the
imgui demo. I don't want to fight against another imgui quirk right now, so will have to live with the loss of a
character when entering text. (This will be the software equivalent of the first stair in a set of stairs being twice
the height of the other stairs; everyone will trip on it pretty much everytime they go to use this feature...)