Imebra_Shim: downgrade warnings about missing metadata.
The DICOM image loader routine is used for CT and RTIMAGEs. The latter does not need to specify image thickness.
Previously, this caused a warning when loading. However, the default logging threshold passes warnings along as
notifications. To reduce noise, since this is expected, the warning has been downgraded in severity.