ContourWholeImages: use direct vertex spacing estimator.
Previously, spacing between vertices for circular FOV ROI contours was picked so that the vertex sampling would
indirectly control the maximum deviation from a perfect circle. This commit adds a more direct estimation method
comparing the geometry of a single adjacent vertex pair either as a right triangle or as an arc. This results in ROI
contours that are more economical in that they satisfy a desired maximum deviation without excessive sampling.