RPC: use TSimpleFileTransport for Drover import/export.
The previous Thrift transport, TFileTransport, is a framed transport with length prefix and 'event' markers interspersed
throughout the emitted file. TSimpleFileTransport is a linear export of the contents without any markers or other
additions. From what I can tell, there is greater cross-language support for TSimpleFileTransport compared with
TFileTransport -- but I;m not certain since I've mostly gathered this info from the mailing list and bug reports which
could be out of date.