Convert hardcoded polyomino call into a plumbing script.
This commit moves from a hard-coded internal op list building pattern to an external script setup. Now the polyomino
operations are parsed from the external script just before being used. This improves testability and makes it easier to
migrate actions when the underlying operations change.

Additional internal actions are planned to be converted to plumbing scripts...