SDL_Viewer: contouring: make undo snapshot on mouse actions.
Previously, copies were made when a mouse unclick action occured. While this is better from a latency perspective, it
results in a missing first entry (i.e., an empty contouring image array) and a duplicate of the current entry. So going
back in history one entry results in effectively the same entry (from the user's perspective).

This commit makes copy entries into the Drover cache when the mouse is first clicked, so the first empty entry is
present, and the final entry is not duplicated.