SDL_Viewer: contouring: relax window focus requirement.
When contouring, there are three windows at play:
  - image viewer
  - image navigation
  - contouring/drawing settings

Because imgui can register a mouse click on a window *before* considering the window to be focused, the prior
mouse/keyboard interception logic would miss first clicks when contouring. This is impactful when contouring because the
undo buffer would not be triggered, so a step would be missing from the undo history.

This commit adds a workaround for detecting when a window has been clicked with a mouse, but is not yet considered
focused. It's sort of a weird edge case, but maybe related to the immediate-mode nature of imgui?