SDL_Viewer: support rudimentary contour/image interpolation.
This implementation uses an extremely simplistic axial interpolation based on image voxel-to-voxel linear interpolation.
Registration is not used and it is assumed the images and contents of the images are already aligned along the axial
direction.

This implementation works *only* for contours that are perfectly aligned on the upper- and lower-most slices.
Though not always accurate for real-world scenarios, it *does* allow for a large number of contours to be added roughly
and then refined manually.