SDL_Viewer: support inline guides via operation argument.
This further improves the utility of using SDL_Viewer in semi-automated workflows where one or more steps need to be
interactive. The guide can be used to walk the user through one or more steps, or even just label the current step so it
is easier for the user to track.