SDL_Viewer: contouring: tweak tanh and gaussian brushes.
The tanh brush needs a little less strength by default, otherwise it produces a rough, jagged contour/boundary. This
commit reduces it from 100% strength to 95% strength.

The gaussian brush seems to be extended slightly too far (3x the radius) for normal use. This commit reduce it from 3x
the brush radius to 2.25x the radius.