SDL_Viewer: contouring: debounce mouse clicks.
Previously, almost all mouse click events would be duplicated. This caused brush strokes to be applied in duplicate,
which has implications for the geometric accuracy of some brushes and sweeps. This commit adds a mouse debounce. The old
behaviour can still be accomplished with repeated mouse clicks, or holding down a mouse button for a small, but
deliberate, amount of time.