DICOM: attempt passthrough of additional top-level UIDs.
This commit aims to simplify addition of UID cross-references between contours and CT images. Previously, many UIDs
which would be referenced at low paranoia levels would be replaced. This commit opts to keep UIDs.

Note that this may cause issues if the UIDs are not valid. Currently, no validity check is performed.