GenerateMapTiles: default to purging backslahes from tmpdir.
Backslashes, present in Windows paths, are also used for escaping characters in operation parameters.
This means that any provided paths (including the default!) either need to escape their backslashes,
or replace them with forwardslashes, which -- from what I can tell -- are also interpretted as valid
paths. This commit adds the latter and a note to the user to remind them about backslashes.