select_directory: reduce to single attempt.
The prior code attempted three times to query the user, which caused issues when the user cancels the selector dialog.
In-band messaging was attempted, but is problematic because it can be hard to differentiate the user cancelling from a
failure to invoke the dialog.

This commit switches to a single attempt and only uses an in-band cancellation message where multiple dialogs are
potentially invoked.