Meshing: use convex hull to find correspondence.
This commit uses the convex hull to help determine the correspondence between upper and lower contours for 1-to-1
polygon stitching. It uses the hull to both seed the stitching, and also selects candidate faces accounting for faces
that appear on the hull.

Note that this method still relies on heuristics when no candidate faces appear on the hull. The heuristics are
greedy-first and do not undergo any specific optimization, so are not optimal.

Also note that the convex hull calculation can sometimes become unstable due to numerical imprecision, which affects the
stitching with sub-optimal faces.