Add MaskParameters operation.
This operation behaves similar to the AdjustVerbosity operation (note: now renamed to MaskVerbosity) in that it presents
a temporary adjustment for children operations. Except it temporarily modifies the global parameter table rather than
the logging mechanism.

A mechanism is provided to enable permanent or transaction semantics so that the modified parameter table replaces the
original if and only if the children operations succeed.

This commit also renames the AdjustVerbosity operation to MaskVerbosity for increased consistency and hopefully slightly
more discoverability.