AnalyzeHistograms: allow all expressions to be saved to parameter table.
This commit adds support for saving both the evaluated quantities (i.e., the LHS of a constraint, e.g., 'Dmean' or
'D(hottest 25%)') and the result of the constraint check (i.e., whether the inequality evaluates to 'true' or 'false').

This functionality further simplifies re-use of this operation with other operations without having to parse and extract
the result from a table or file.