ModelIVIM: rename 'auc' to 'auc-simple'.
An alternative (arguably better) method for extracting the AUC would be to fit with a model, and then integrate it over
known bounds (e.g., [0,inf)) which would give an AUC estimate that is less sensitive to the specific b-values used in a
given study. This technically might involve extrapolation and the model could be more complicated, so the present,
direct-integrate model was renamed to 'auc-simple' to reflect.