AnalyzeHistograms: accept arbitrary column headings.
This commit allows the user to provide arbitrary column headings for individual metadata. This reduces the need to
'pack' variables into the UserComment or Description columns and instead makes it easier to retrieve/process records
with a fixed columnar layout.

Metadata are retrieved from the current line sample metadata first, and the global parameter table second. Failing that,
they are left empty so the layout of the table is not impacted when specific metadata are not available.