Metadata: inject_metadata: disable self-expansion.
There is high potential for infinite recursion. While this commit does limit what the user can accomplish using metadata
replacement (i.e., complicated logic based on nested macro replacments), it significantly reduces the risk of accidental
recursion when expected metadata are not present.

Note that the global parameter table is still available and expanded preferentially when key-values are provided with
the '${...}' notation, so it is still possible to support macro programs with runtime indirection...