ConvertParametersToTable: emit 'wide' tables.
Adding all matching key selections to a single row makes it easier to merge multiple records consistently, e.g., 'group
by' in SQL or joins. Previously, the structure was essentially a list of key-values with each key-value pair printed one
per row. While this structure simplifies grepping and re-insertion, it can make it impossible to determine which rows
correspond to a specific entity.

Providing an option to emit either 'wide' or 'tall' format may be provided in the future.