Added operation filtering and tagging system.
This commit adds a tgging system to the operations. The idea is to use it for both explicit tagging (i.e., labeling each
operation with metadata) and runtime-generated tagging to better filter the list of operations.

A small number of operations have been tagged for now, just to debug the tagging and filtering system. Additional tags
are forthcoming.