Create regex_group class.
This class is used to deal with the situation where there are multiple regexes used to control which methods or options
are performed. The laborious process of escaping optional characters is now automated, including the situation where
there are conflicting mandatory characters -- which are handled by gradually maiing optional characters mandatory.